requirements.txt
fastapi
uvicorn
python-dotenv
elasticsearch
httpx

